requirements.txt
torch
torchvision
transformers
Pillow
numpy
pandas
opencv-python-headless
ultralytics
faiss-cpu
requests
tqdm
openpyxl
scikit-learn
python-dotenv
